backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
